fix: 修复临时会话 ttl 类型错误

+++ LeanCloud.Realtime/Internal/Command/ConversationCommand.cs
@@ -53,7 +53,7 @@ namespace LeanCloud.Realtime.Internal
 
         public ConversationCommand TempConvTTL(double tempConvTTL)
         {
-            return new ConversationCommand(this.Argument("tempConvTTL", tempConvTTL));
+            return new ConversationCommand(this.Argument("tempConvTTL", (long)tempConvTTL));
         }
 
         public ConversationCommand Attr(IDictionary<string, object> attr)
